Removed ft_run_game
